feat: refactor email sending logic and update email templates for dynamic content
