+++ src/constants/index.ts
@@ -2,3 +2,4 @@ export * from './messages.ts'
 export * from './fileLimits.ts'
 export * from './pagination.ts'
 export * from './queues.ts'
+export * from './mailer.ts'
