+++ src/libs/index.ts
@@ -1,2 +1,2 @@
-export { sendEmail, type SendEmailProps } from './nodemailer'
+export { sendEmail } from './nodemailer'
 export { logtailServer as logger, logtailWorker as logWorker } from './logtail'
