+++ src/libs/nodemailer.ts
@@ -1,14 +1,10 @@
 import nodemailer from 'nodemailer'
 import { env } from '../config'
-import {
-  verificationEmail,
-  passwordResetEmail,
-  orderConfirmationEmail,
-} from './emailTemplates'
-import { userMessage, type EmailJobType } from '../constants'
-import type { Order } from '../types'
+import { getEmailHtml } from '../utils'
+import { attachments, subjectMap, userMessage } from '../constants'
+import type { SendEmailProps } from '../types'
 
-const transporter = nodemailer.createTransport({
+const transportOptions = {
   service: 'gmail',
   host: 'smtp.gmail.com',
   port: 465,
@@ -17,89 +13,21 @@ const transporter = nodemailer.createTransport({
     user: env.mailerUser,
     pass: env.mailerPass,
   },
-})
+}
 
-export type SendEmailProps =
-  | {
-      type: Extract<EmailJobType, 'orderConfirmation'>
-      toAddress: string
-      toName: string
-      order: Order
-    }
-  | {
-      type: Extract<EmailJobType, 'verification' | 'passwordReset'>
-      toAddress: string
-      toName: string
-      tokenLink: string
-    }
+const transporter = nodemailer.createTransport(transportOptions)
 
 export async function sendEmail(props: SendEmailProps) {
-  const baseLink = env.clientBaseUrl!
-  const { type, toAddress, toName } = props
-
-  const subject = {
-    verification: 'Book Shop - Verify your email address',
-    passwordReset: 'Book Shop - Forgotten Password',
-    orderConfirmation: 'Book Shop - Order Confirmation',
-  }[type]
-
-  const logo = 'bookshop-logo-cid'
-
-  const getHtml = () => {
-    if (type === 'orderConfirmation') {
-      const { order } = props
-      return orderConfirmationEmail({
-        toName: toName,
-        orderNumber: order.paymentId.slice(-6).toUpperCase(),
-        items: order.items.map(({ title, quantity, price, discount }) => ({
-          title,
-          quantity,
-          price,
-          discount,
-        })),
-        total: order.total,
-        currency: order.currency,
-        address: [
-          order.address?.line1,
-          order.address?.line2,
-          order.address?.city,
-          order.address?.state,
-          order.address?.country,
-        ]
-          .filter(Boolean)
-          .join(', '),
-      }).html
-    } else if (type === 'verification') {
-      const { toName, tokenLink } = props
-      return verificationEmail({ toName, tokenLink, baseLink, logo }).html
-    } else if (type === 'passwordReset') {
-      const { toName, tokenLink } = props
-      return passwordResetEmail({ toName, tokenLink, baseLink, logo }).html
-    } else {
-      throw new Error('Invalid email type')
-    }
-  }
-
-  const html = getHtml()
-
   const mailOptions = {
     from: env.mailerUser,
-    to: toAddress,
-    subject,
-    html,
-    attachments: [
-      {
-        filename: 'bookshop-logo.png',
-        path: process.cwd() + '/src/assets/images/bookshop-logo.png',
-        contentType: 'image/png',
-        cid: logo,
-      },
-    ],
+    to: props.toAddress,
+    subject: subjectMap[props.type],
+    html: getEmailHtml(props),
+    attachments,
   }
 
   try {
-    const mailSent = await transporter.sendMail(mailOptions)
-    return mailSent
+    return await transporter.sendMail(mailOptions)
   } catch (error) {
     throw new Error(userMessage.sendEmail, { cause: error })
   }
