+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './authors.t'
 export * from './news.t'
 export * from './users.t'
 export * from './orders.t'
+export * from './worker.t'
