+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export { formatUptime } from './formatUptime'
 export { verifyJWTAccess, verifyJWTRefresh } from './verifyJWT'
 export { signAccessToken, signRefreshToken } from './signJWT'
 export { uploadFile } from './uploadFile'
+export { getEmailHtml } from './getEmailHtml'
